Move new types to common
