+++ datafusion/proto/src/physical_plan/to_proto.rs
@@ -45,19 +45,11 @@ use datafusion::{
     },
     physical_plan::expressions::LikeExpr,
 };
-use datafusion_common::config::{
-    ColumnOptions, CsvOptions, FormatOptions, JsonOptions, ParquetOptions,
-    TableParquetOptions,
-};
+use datafusion_common::config::FormatOptions;
 use datafusion_common::{
-    file_options::{csv_writer::CsvWriterOptions, json_writer::JsonWriterOptions},
-    internal_err, not_impl_err,
-    parsers::CompressionTypeVariant,
-    stats::Precision,
-    DataFusionError, JoinSide, Result,
+    internal_err, not_impl_err, stats::Precision, DataFusionError, JoinSide, Result,
 };
 
-use crate::logical_plan::csv_writer_options_to_proto;
 use crate::protobuf::{
     self, copy_to_node, physical_aggregate_expr_node, physical_window_expr_node,
     ArrowOptions, AvroOptions, PhysicalSortExprNode, PhysicalSortExprNodeCollection,
@@ -836,167 +828,6 @@ impl TryFrom<&FileSinkConfig> for protobuf::FileSinkConfig {
             output_schema: Some(conf.output_schema.as_ref().try_into()?),
             table_partition_cols,
             overwrite: conf.overwrite,
-        })
-    }
-}
-
-impl From<&CompressionTypeVariant> for protobuf::CompressionTypeVariant {
-    fn from(value: &CompressionTypeVariant) -> Self {
-        match value {
-            CompressionTypeVariant::GZIP => Self::Gzip,
-            CompressionTypeVariant::BZIP2 => Self::Bzip2,
-            CompressionTypeVariant::XZ => Self::Xz,
-            CompressionTypeVariant::ZSTD => Self::Zstd,
-            CompressionTypeVariant::UNCOMPRESSED => Self::Uncompressed,
-        }
-    }
-}
-
-impl TryFrom<&CsvWriterOptions> for protobuf::CsvWriterOptions {
-    type Error = DataFusionError;
-
-    fn try_from(opts: &CsvWriterOptions) -> Result<Self, Self::Error> {
-        Ok(csv_writer_options_to_proto(
-            &opts.writer_options,
-            &opts.compression,
-        ))
-    }
-}
-
-impl TryFrom<&JsonWriterOptions> for protobuf::JsonWriterOptions {
-    type Error = DataFusionError;
-
-    fn try_from(opts: &JsonWriterOptions) -> Result<Self, Self::Error> {
-        let compression: protobuf::CompressionTypeVariant = opts.compression.into();
-        Ok(protobuf::JsonWriterOptions {
-            compression: compression.into(),
-        })
-    }
-}
-
-impl TryFrom<&ParquetOptions> for protobuf::ParquetOptions {
-    type Error = DataFusionError;
-
-    fn try_from(value: &ParquetOptions) -> Result<Self, Self::Error> {
-        Ok(protobuf::ParquetOptions {
-            enable_page_index: value.enable_page_index,
-            pruning: value.pruning,
-            skip_metadata: value.skip_metadata,
-            metadata_size_hint_opt: value.metadata_size_hint.map(|v| protobuf::parquet_options::MetadataSizeHintOpt::MetadataSizeHint(v as u64)),
-            pushdown_filters: value.pushdown_filters,
-            reorder_filters: value.reorder_filters,
-            data_pagesize_limit: value.data_pagesize_limit as u64,
-            write_batch_size: value.write_batch_size as u64,
-            writer_version: value.writer_version.clone(),
-            compression_opt: value.compression.clone().map(protobuf::parquet_options::CompressionOpt::Compression),
-            dictionary_enabled_opt: value.dictionary_enabled.map(protobuf::parquet_options::DictionaryEnabledOpt::DictionaryEnabled),
-            dictionary_page_size_limit: value.dictionary_page_size_limit as u64,
-            statistics_enabled_opt: value.statistics_enabled.clone().map(protobuf::parquet_options::StatisticsEnabledOpt::StatisticsEnabled),
-            max_statistics_size_opt: value.max_statistics_size.map(|v| protobuf::parquet_options::MaxStatisticsSizeOpt::MaxStatisticsSize(v as u64)),
-            max_row_group_size: value.max_row_group_size as u64,
-            created_by: value.created_by.clone(),
-            column_index_truncate_length_opt: value.column_index_truncate_length.map(|v| protobuf::parquet_options::ColumnIndexTruncateLengthOpt::ColumnIndexTruncateLength(v as u64)),
-            data_page_row_count_limit: value.data_page_row_count_limit as u64,
-            encoding_opt: value.encoding.clone().map(protobuf::parquet_options::EncodingOpt::Encoding),
-            bloom_filter_on_read: value.bloom_filter_on_read,
-            bloom_filter_on_write: value.bloom_filter_on_write,
-            bloom_filter_fpp_opt: value.bloom_filter_fpp.map(protobuf::parquet_options::BloomFilterFppOpt::BloomFilterFpp),
-            bloom_filter_ndv_opt: value.bloom_filter_ndv.map(protobuf::parquet_options::BloomFilterNdvOpt::BloomFilterNdv),
-            allow_single_file_parallelism: value.allow_single_file_parallelism,
-            maximum_parallel_row_group_writers: value.maximum_parallel_row_group_writers as u64,
-            maximum_buffered_record_batches_per_stream: value.maximum_buffered_record_batches_per_stream as u64,
-        })
-    }
-}
-
-impl TryFrom<&ColumnOptions> for protobuf::ColumnOptions {
-    type Error = DataFusionError;
-
-    fn try_from(value: &ColumnOptions) -> Result<Self, Self::Error> {
-        Ok(protobuf::ColumnOptions {
-            compression_opt: value
-                .compression
-                .clone()
-                .map(protobuf::column_options::CompressionOpt::Compression),
-            dictionary_enabled_opt: value
-                .dictionary_enabled
-                .map(protobuf::column_options::DictionaryEnabledOpt::DictionaryEnabled),
-            statistics_enabled_opt: value
-                .statistics_enabled
-                .clone()
-                .map(protobuf::column_options::StatisticsEnabledOpt::StatisticsEnabled),
-            max_statistics_size_opt: value.max_statistics_size.map(|v| {
-                protobuf::column_options::MaxStatisticsSizeOpt::MaxStatisticsSize(
-                    v as u32,
-                )
-            }),
-            encoding_opt: value
-                .encoding
-                .clone()
-                .map(protobuf::column_options::EncodingOpt::Encoding),
-            bloom_filter_enabled_opt: value
-                .bloom_filter_enabled
-                .map(protobuf::column_options::BloomFilterEnabledOpt::BloomFilterEnabled),
-            bloom_filter_fpp_opt: value
-                .bloom_filter_fpp
-                .map(protobuf::column_options::BloomFilterFppOpt::BloomFilterFpp),
-            bloom_filter_ndv_opt: value
-                .bloom_filter_ndv
-                .map(protobuf::column_options::BloomFilterNdvOpt::BloomFilterNdv),
-        })
-    }
-}
-
-impl TryFrom<&TableParquetOptions> for protobuf::TableParquetOptions {
-    type Error = DataFusionError;
-    fn try_from(value: &TableParquetOptions) -> Result<Self, Self::Error> {
-        let column_specific_options = value
-            .column_specific_options
-            .iter()
-            .map(|(k, v)| {
-                Ok(protobuf::ColumnSpecificOptions {
-                    column_name: k.into(),
-                    options: Some(v.try_into()?),
-                })
-            })
-            .collect::<Result<Vec<_>>>()?;
-        Ok(protobuf::TableParquetOptions {
-            global: Some((&value.global).try_into()?),
-            column_specific_options,
-        })
-    }
-}
-
-impl TryFrom<&CsvOptions> for protobuf::CsvOptions {
-    type Error = DataFusionError; // Define or use an appropriate error type
-
-    fn try_from(opts: &CsvOptions) -> Result<Self, Self::Error> {
-        let compression: protobuf::CompressionTypeVariant = opts.compression.into();
-        Ok(protobuf::CsvOptions {
-            has_header: opts.has_header.map_or_else(Vec::new, |h| vec![h as u8]),
-            delimiter: vec![opts.delimiter],
-            quote: vec![opts.quote],
-            escape: opts.escape.map_or_else(Vec::new, |e| vec![e]),
-            compression: compression.into(),
-            schema_infer_max_rec: opts.schema_infer_max_rec as u64,
-            date_format: opts.date_format.clone().unwrap_or_default(),
-            datetime_format: opts.datetime_format.clone().unwrap_or_default(),
-            timestamp_format: opts.timestamp_format.clone().unwrap_or_default(),
-            timestamp_tz_format: opts.timestamp_tz_format.clone().unwrap_or_default(),
-            time_format: opts.time_format.clone().unwrap_or_default(),
-            null_value: opts.null_value.clone().unwrap_or_default(),
-        })
-    }
-}
-
-impl TryFrom<&JsonOptions> for protobuf::JsonOptions {
-    type Error = DataFusionError;
-
-    fn try_from(opts: &JsonOptions) -> Result<Self, Self::Error> {
-        let compression: protobuf::CompressionTypeVariant = opts.compression.into();
-        Ok(protobuf::JsonOptions {
-            compression: compression.into(),
-            schema_infer_max_rec: opts.schema_infer_max_rec as u64,
         })
     }
 }
